Move methods from DawnCompiler to CodeGen Driver
