+++ dawn/src/dawn/Compiler/DawnCompiler.cpp
@@ -450,41 +450,4 @@ run(const std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>&
   return compiler.optimize(stencilInstantiationMap, groups);
 }
 
-namespace codegen {
-
-codegen::Backend parseBackendString(const std::string& backendStr) {
-  if(backendStr == "gt" || backendStr == "gridtools") {
-    return codegen::Backend::GridTools;
-  } else if(backendStr == "naive" || backendStr == "cxxnaive" || backendStr == "c++-naive") {
-    return codegen::Backend::CXXNaive;
-  } else if(backendStr == "ico" || backendStr == "naive-ico" || backendStr == "c++-naive-ico") {
-    return codegen::Backend::CXXNaiveIco;
-  } else if(backendStr == "cuda" || backendStr == "CUDA") {
-    return codegen::Backend::CUDA;
-  } else {
-    throw CompileError("Backend not supported");
-  }
-}
-
-std::unique_ptr<TranslationUnit>
-run(const std::map<std::string, std::shared_ptr<iir::StencilInstantiation>>& context,
-    Backend backend, const Options& options) {
-  switch(backend) {
-  case Backend::CUDA:
-    return cuda::run(context, options);
-  case Backend::CXXNaive:
-    return cxxnaive::run(context, options);
-  case Backend::CXXNaiveIco:
-    return cxxnaiveico::run(context, options);
-  case Backend::GridTools:
-    return gt::run(context, options);
-  case Backend::CXXOpt:
-    throw std::invalid_argument("Backend not supported");
-  }
-  // This line should not be needed but the compiler seems to complain if it is not present.
-  return nullptr;
-}
-
-} // namespace codegen
-
 } // namespace dawn
